Добавил простое логирование, запрет на удаление тегов, если они закреплены за вопросами, вопросы больше не возвращаются с ревью, чтобы получить ревью, нужно запросить конкретный вопрос по айди. Теперь вопрос нельзя изменить, если его автор - не ты. По неверным айдишникам поиск теперь возвращает null (пока не придумал, как сделать красивше)

+++ src/main/java/com/example/blps_2/service/TagService.java
@@ -24,7 +24,7 @@ public class TagService {
     }
 
     public Tag findById(long id) {
-        return tagRepository.findById(id).orElse(new Tag());
+        return tagRepository.findById(id).orElse(null);
     }
 
     @Transactional
@@ -42,9 +42,13 @@ public class TagService {
     }
 
     @Transactional
-    public void deleteById(long id) {
+    public String deleteById(long id) {
         Tag tag = tagRepository.findById(id).orElse(new Tag());
+        if (tag.getQuestions() != null) {
+            return "This tag is in use. No delete can be provided.";
+        }
         tagRepository.deleteById(id);
+        return "Success";
     }
 
 }
